+++ examples/app/app.controller.ts
@@ -1,4 +1,4 @@
-import {LoggerService} from "../../src/index";
+import {LoggerService, MysqlService} from "../../src/index";
 import {HttpService} from "../../src/index";
 import {Controller, Inject, Route} from "../../src/index";
 
@@ -7,20 +7,23 @@ import {Controller, Inject, Route} from "../../src/index";
 export default class AppController extends Controller
 {
   constructor(public http: HttpService,
-              public logger: LoggerService)
+              public logger: LoggerService,
+              public mysql: MysqlService)
   {
     super(http);
   }
 
 
   @Route('/bar')
-  bar(params: any)
+  async bar(params: any)
   {
     const log = this.logger.action('AppController.bar');
 
     log.info('bar', params.name, params.telephone);
 
-    return 'bar';
+    const results = await this.mysql.query('SELECT 1+2 as test');
+
+    return `${results[0].test}`;
   }