+++ examples/app/app.module.ts
@@ -1,4 +1,4 @@
-import {Module} from "../../src/index";
+import {Module, MysqlService} from "../../src/index";
 import {LoggerModule} from "../../src/index";
 import {HttpService} from "../../src/index";
 import AppController from "./app.controller";
@@ -11,11 +11,18 @@ export default class AppModule extends Module
     return [
       LoggerModule.config(options =>
       {
-        options.level = 'debug';
+        options.level = 'silly';
       }),
       HttpService.config(options =>
       {
         options.port = 3000;
+      }),
+      MysqlService.config(options =>
+      {
+        options.host = "localhost";
+        options.user = "test.user";
+        options.password = "123456";
+        options.database = "ready_db";
       }),
       AppController,
     ];